requirements-prod.txt
Django==5.2.7
djangorestframework
python-decouple
Pillow
django-crispy-forms
crispy-bootstrap5
psycopg2-binary
gunicorn
whitenoise
django-cors-headers
django-filter
